query refactoring: break up into query, parse, and eval
